+++ crates/alpaca-trade/src/account/client.rs
@@ -1,8 +1,10 @@
 use std::sync::Arc;
 
+use crate::account::Account;
 use crate::client::Inner;
+use crate::error::Error;
+use crate::transport::endpoint::Endpoint;
 
-#[allow(dead_code)]
 #[derive(Debug, Clone)]
 pub struct AccountClient {
     inner: Arc<Inner>,
@@ -11,5 +13,17 @@ pub struct AccountClient {
 impl AccountClient {
     pub(crate) fn new(inner: Arc<Inner>) -> Self {
         Self { inner }
+    }
+
+    pub async fn get(&self) -> Result<Account, Error> {
+        self.inner
+            .http
+            .get_json(
+                &self.inner.base_url,
+                Endpoint::Account,
+                &self.inner.auth,
+                vec![],
+            )
+            .await
     }
 }